requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 2.0.2
Flask_SQLAlchemy == 2.5.1
Werkzeug == 2.0.2
nanoid == 2.0.0
